[Spyll, Oyl3d] Move refelection impl classes to Oyl3d
Each project should use it's own implementation for using reflection
info, and each project will consume the spyll descriptors for that
implementation
