+++ Source/Oyl3D/Core/Core/Reflection/Assembly.cpp
@@ -1,62 +1,86 @@
-#include "pch.h"
 #include "Assembly.h"
 
-#include "Core/Logging/Logging.h"
+//extern "C"
+//{
+	//void
+	//_PopulateReflectionAssembly(Spyll::Reflection::Assembly* a_assembly)
+	//{
+	//	using namespace Spyll::Reflection;
 
-namespace Oyl::Reflection
+	//	Type type;
+	//	type.m_assembly = a_assembly;
+	//}
+//}
+
+namespace Spyll::Reflection
 {
-	bool
-	Assembly::RegisterType(Type* a_type)
+	std::vector<Assembly*> Assembly::s_assemblies;
+	std::unordered_map<std::string_view, Assembly*> Assembly::s_assemblyMap;
+
+	Assembly::Assembly() {}
+
+	Assembly::~Assembly() {}
+
+	const std::vector<Type>&
+	Assembly::GetTypes() const
 	{
-		for (auto& assemblyPtr : Assemblies())
-		{
-			auto& ownedTypes = assemblyPtr->m_ownedTypes;
-			auto iter = std::find_if(
-				ownedTypes.begin(),
-				ownedTypes.end(),
-				[a_type](const Type* a_value)
-				{
-					return *a_value == *a_type;
-				}
-			);
-
-			if (iter != ownedTypes.end())
-			{
-				OYL_LOG_ERROR(
-					"Type \"{}\" trying to be registered under assembly \"{}\" is already registered to assembly \"{}\"",
-					a_type->FullName(),
-					m_name,
-					assemblyPtr->m_name
-				);
-				return false;
-			}
-		}
-
-		m_ownedTypes.push_back(a_type);
-		a_type->m_assembly = this;
-
-		return true;
+		return m_types;
+	}
+	
+	const std::vector<Function>&
+	Assembly::GetFreeFunctions() const
+	{
+		return m_functions;
 	}
 
-	bool
-	Assembly::UnRegisterType(Type* a_type)
+	const Type*
+	Assembly::GetType(std::string_view a_typeName) const
 	{
-		auto iter = std::find_if(
-			m_ownedTypes.begin(),
-			m_ownedTypes.end(),
-			[a_type](const auto& a_value)
-			{
-				return *a_value == *a_type;
-			}
-		);
+		auto iter = m_typeMap.find(a_typeName);
+		if (iter == m_typeMap.end())
+			return nullptr;
+		return iter->second;
+	}
 
-		if (iter == m_ownedTypes.end())
+	const Function*
+	Assembly::GetFunction(std::string_view a_functionName) const
+	{
+		auto iter = m_functionMap.find(a_functionName);
+		if (iter == m_functionMap.end())
+			return nullptr;
+		return iter->second;
+	}
+
+	const std::vector<Assembly*>&
+	Assembly::GetAssemblies()
+	{
+		return s_assemblies;
+	}
+
+	const Assembly*
+	Assembly::GetAssembly(std::string_view a_assemblyName)
+	{
+		auto iter = s_assemblyMap.find(a_assemblyName);
+		if (iter == s_assemblyMap.end())
+			return nullptr;
+		return iter->second;
+	}
+
+	void
+	Assembly::SetupAssemblyMaps()
+	{
+		m_typeMap.reserve(m_types.size());
+		for (size_t i = 0; i < m_types.size(); i++)
 		{
-			return false;
+			auto& type = m_types[i];
+			m_typeMap.emplace(type.GetQualifiedName(), &type);
 		}
 
-		m_ownedTypes.erase(iter);
-		a_type->m_assembly = nullptr;
-		return true;
+		m_functionMap.reserve(m_functions.size());
+		for (size_t i = 0; i < m_functions.size(); i++)
+		{
+			auto& function = m_functions[i];
+			m_functionMap.emplace(function.GetQualifiedName(), &function);
+		}
 	}
 }
