+++ Source/Oyl3D/Core/Core/Reflection/MemberInfo.h
