+++ Source/Oyl3D/Core/Core/Reflection/NamedDeclaration.h
