+++ Source/Oyl3D/Core/Core/Reflection/Type.cpp
@@ -1,166 +1,74 @@
-#include "pch.h"
 #include "Type.h"
 
-namespace Oyl::Reflection
+#include "Assembly.h"
+
+namespace Spyll::Reflection
 {
-	Type::~Type()
+	Type::Type() {}
+
+	Type::~Type() {}
+
+	TypeId
+	Type::GetTypeId() const
 	{
-#	ifdef _OYL_TYPE_HEAP_ALLOCATED_FACTORY
-		delete m_factory;
-#	else
-		Factory()->~GenericFactory();
-		std::memset(&m_factoryData, 0, sizeof(m_factoryData));
-#	endif
+		return m_typeId;
+	}
+
+	std::string_view
+	Type::GetName() const
+	{
+		return m_name;
+	}
+
+	std::string_view
+	Type::GetQualifiedName() const
+	{
+		return m_qualifiedName;
+	}
+
+	uint32_t
+	Type::GetSize() const
+	{
+		return m_size;
 	}
 
 	bool
-	Type::IsConvertibleTo(Oyl::TypeId a_typeId) const
+	Type::IsPrimitive() const
 	{
-		if (a_typeId == TypeId::Null)
-		{
-			return false;
-		}
-
-		if (m_typeId == a_typeId)
-		{
-			return true;
-		}
-
-		if (Type* parent = TryGet(m_baseTypeId))
-		{
-			return parent->IsConvertibleTo(a_typeId);
-		}
-
-		return false;
+		return m_isPrimitive;
 	}
 
-	bool
-	Type::IsConvertibleTo(const Type* a_type) const
+	const std::vector<Type*>&
+	Type::GetBaseTypes() const
 	{
-		return IsConvertibleTo(a_type->TypeId());
+		return m_baseTypes;
 	}
 
-	bool
-	Type::IsConvertibleFrom(Oyl::TypeId a_typeId) const
+	const std::vector<Field*>&
+	Type::GetFields() const
 	{
-		if (a_typeId == TypeId::Null)
-		{
-			return false;
-		}
-
-		if (m_typeId == a_typeId)
-		{
-			return true;
-		}
-
-		// Check if the other type is convertible to this type
-
-		if (Type* otherType = TryGet(a_typeId))
-		{
-			return otherType->IsConvertibleTo(m_typeId);
-		}
-
-		return false;
+		return m_fields;
 	}
 
-	bool
-	Type::IsConvertibleFrom(const Type* a_type) const
+	const std::vector<Function*>&
+	Type::GetFunctions() const
 	{
-		return IsConvertibleFrom(a_type->TypeId());
+		return m_functions;
 	}
 
-	Type* Type::TryGet(Oyl::TypeId a_id) noexcept
+	const Type*
+	Type::Get(std::string_view a_name)
 	{
-		Type* result = nullptr;
-		auto& types  = Types();
-		if (auto iter = types.find(a_id); iter != types.end())
-		{
-			result = iter->second;
-		}
+		auto& assemblies = Assembly::GetAssemblies();
 
-		return result;
-	}
-
-	Type* Type::TryGet(std::string_view a_fullyQualifiedTypeName) {
-		Type* result = nullptr;
-		auto& types  = Types();
-		for (auto& [typeId, type] : types)
+		for (const Assembly* assembly : assemblies)
 		{
-			if (a_fullyQualifiedTypeName.compare(type->FullName()) == 0)
+			if (const Type* type = assembly->GetType(a_name))
 			{
-				result = type;
-				break;
+				return type;
 			}
 		}
-		return result;
-	}
 
-	void Type::ProcessName(const std::type_info& a_typeInfo) {
-		const char* typeName = a_typeInfo.name();
-
-		// type_info.name() is implementation specific
-		#if defined _MSC_BUILD
-		// for user types, MSVC uses the format [class, struct] [<namespace>::]<classname> [cv and pointer status]
-		// ie. "Type const*" is "class Next::Reflection::Type const * __ptr64" in x86_64
-		const char* classString     = strstr(typeName, "class");
-		const char* structString    = strstr(typeName, "struct");
-		const char* firstWhitespace = strchr(typeName, ' ');
-			
-		if (!(classString || structString))
-		{
-			std::ptrdiff_t difference;
-
-			// Assume this is a primitive type
-			if (firstWhitespace)
-			{
-				// Oh wow, const char* - const char* = long long, who knew?
-				difference = (firstWhitespace + 1) - typeName;
-			} else
-			{
-				difference = strlen(typeName);
-			}
-				
-			m_name     = std::string(typeName, difference);
-			m_fullName = m_name;
-
-			return;
-		}
-
-		// This probably shouldn't be an assert but im lazy
-		// There should always be whitespace in a user-defined-class name
-		OYL_ASSERT(firstWhitespace);
-
-		const char* fullNameStart = firstWhitespace + 1;
-
-		const char* shortNameStart = fullNameStart;
-
-		// Find the last instance of the scope operator
-		while (true)
-		{
-			const char* find = strstr(shortNameStart, "::");
-			if (!find)
-			{
-				break;
-			}
-			shortNameStart = find + 2;
-		}
-
-		const char* nextWhitespace = strchr(shortNameStart, ' ');
-
-		std::ptrdiff_t fullNameLength;
-		std::ptrdiff_t shortNameLength;
-		if (nextWhitespace)
-		{
-			fullNameLength  = nextWhitespace - fullNameStart;
-			shortNameLength = nextWhitespace - shortNameStart;
-		} else
-		{
-			fullNameLength  = strlen(fullNameStart);
-			shortNameLength = strlen(shortNameStart);
-		}
-			
-		m_name     = std::string(shortNameStart, shortNameLength);
-		m_fullName = std::string(fullNameStart, fullNameLength);
-		#endif
+		return nullptr;
 	}
 }
