+++ Source/Oyl3D/Core/Core/Reflection/Variable.h
