+++ Source/Spyll/Spyll/ClangTool/ReflectionGenerator.h
@@ -2,8 +2,8 @@
 
 #include <clang/AST/DeclCXX.h>
 
-#include "Spyll/Spyll/Descriptors/DescriptorEnums.h"
-#include "Spyll/Spyll/Descriptors/ReflectionDescriptor.h"
+#include "Spyll/Spyll/DescriptorEnums.h"
+#include "Spyll/Spyll/ReflectionDescriptor.h"
 
 namespace Spyll
 {
