+++ Source/Spyll/Spyll/Spyll/DescriptorSerialization.h
