--- Source/Spyll/Spyll/Spyll/TypeDatabase.cpp
@@ -1,20 +1,0 @@
-#include "TypeDatabase.h"
-
-namespace Spyll::Reflection
-{
-	struct TypeDatabase::Impl
-	{
-		
-	};
-	
-	TypeDatabase::TypeDatabase()
-	{
-		m_impl = new Impl;
-	}
-
-	TypeDatabase::~TypeDatabase()
-	{
-		delete m_impl;
-		m_impl = nullptr;
-	}
-}
