--- Source/Spyll/Spyll/Spyll/TypeDatabase.h
@@ -1,15 +1,0 @@
-#pragma once
-
-namespace Spyll::Reflection
-{
-	class TypeDatabase final
-	{
-	public:
-		TypeDatabase();
-
-		~TypeDatabase();
-	private:
-		struct Impl;
-		Impl* m_impl;
-	};
-}
