+++ Source/Spyll/Spyll/Spyll/TypeDescriptor.cpp
