+++ Source/Spyll/Spyll/Spyll/VariableDescriptor.cpp
