+++ Source/Spyll/Spyll/Spyll/VariableDescriptor.h
